new test with strand 1 -1
