+++ pygenometracks/tests/test_bed_and_gtf_tracks.py
@@ -461,28 +461,6 @@ display = collapsed
 with open(os.path.join(ROOT, "bed_genes_rgb.ini"), 'w') as fh:
     fh.write(browser_tracks)
 
-browser_tracks = """
-[macs2 broadPeak]
-file = broadPeak.broadPeak
-title = broadPeak
-file_type = bed
-
-[spacer]
-
-[macs2 gappedPeak]
-file = gappedPeak.gappedPeak
-title = gappedPeak
-file_type = bed
-
-[spacer]
-
-[macs2 filteredbed]
-file = filtered.results.bed
-title = filtered.results.bed (strange format)
-file_type = bed
-"""
-with open(os.path.join(ROOT, "bed_unusual_formats.ini"), 'w') as fh:
-    fh.write(browser_tracks)
 
 wrong_track = """
 [test gtf]
@@ -745,24 +723,6 @@ def test_plot_tracks_bed_all_label_inside_Xdec():
     args = f"--tracks {ini_file} --region {region} "\
            "--trackLabelFraction 0.2 --width 38 --dpi 130 "\
            "--trackLabelHAlign right --decreasingXAxis "\
-           f"--outFileName {outfile.name}".split()
-    pygenometracks.plotTracks.main(args)
-    res = compare_images(expected_file,
-                         outfile.name, tolerance)
-    assert res is None, res
-
-    os.remove(outfile.name)
-
-
-def test_plot_tracks_bed_unusual_format():
-
-    outfile = NamedTemporaryFile(suffix='.png', prefix='pyGenomeTracks_test_',
-                                 delete=False)
-    ini_file = os.path.join(ROOT, "bed_unusual_formats.ini")
-    region = "X:20000-40000"
-    expected_file = os.path.join(ROOT, 'master_bed_unusual_formats.png')
-    args = f"--tracks {ini_file} --region {region} "\
-           "--trackLabelFraction 0.2 --width 38 --dpi 130 "\
            f"--outFileName {outfile.name}".split()
     pygenometracks.plotTracks.main(args)
     res = compare_images(expected_file,
