+++ pygenometracks/tests/test_data/strange_strand.bed
@@ -1,0 +1,3 @@
+chr1	0	100	pos	0	1
+chr1	200	300	neg	0	-1
+chr1	350	400	no_strand	0	.
