+++ pygenometracks/tests/test_data/strange_strand.ini
@@ -1,0 +1,3 @@
+
+[strange_strand]
+file = strange_strand.bed
